// src/types.ts
export enum LoadingState {
    NONE = 0,
    LAODING = 1,
    LOADED = 2,
    ERROR = 2,
}

export enum MessageType {
    MESSAGE_TYPE_ONLINE = 2,
    MESSAGE_TYPE_RELOAD = 3,
    MESSAGE_TYPE_SCORE = 8,
    MESSAGE_TYPE_GIFT_LINK = 10,
    MESSAGE_TYPE_DEADLINE = 11,
    MESSAGE_TYPE_BATCH = 12,
    V = 10,
}